4.21

+++ part4/blog-list-backend/controllers/blogs.js
@@ -41,12 +41,24 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-  try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
-  } catch (error) {
-    next(error)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
   }
+  const user = await User.findById(decodedToken.id)
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === user._id.toString()) {
+    try {
+      await Blog.findByIdAndRemove(request.params.id)
+      user.blogs = user.blogs.filter(blog => blog.toString() !== request.params.id) // removes the blog from users blogs array
+      await user.save()
+      response.status(204).end()
+    } catch (error) {
+      next(error)
+    }
+  } else { return response.status(401).json({ error: 'unauthorized user' }) }
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
